+++ src/base/ValidatesTokens.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace venveo\oauthclient\base;
+
+use venveo\oauthclient\models\Token;
+
+/**
+ * @package venveo\oauthclient\base
+ */
+interface ValidatesTokens
+{
+    public static function checkToken(Token $token): bool;
+}