+++ src/models/App.php
@@ -8,6 +8,8 @@ use craft\validators\UniqueValidator;
 use venveo\oauthclient\base\Provider;
 use venveo\oauthclient\Plugin;
 use venveo\oauthclient\records\App as AppRecord;
+use venveo\oauthclient\records\Token as TokenRecord;
+use craft\elements\User;
 
 /**
  * Class App
@@ -47,6 +49,7 @@ class App extends Model
 
     /**
      * Parse any environment variables and return the client ID
+     *
      * @return string
      */
     public function getClientId(): string
@@ -56,6 +59,7 @@ class App extends Model
 
     /**
      * Parse any environment variables and return the client secret
+     *
      * @return string
      */
     public function getClientSecret(): string
@@ -65,6 +69,7 @@ class App extends Model
 
     /**
      * Get the scopes for the app
+     *
      * @param bool $forTable If true, we'll format the output for Craft's table field
      * @return array
      */
@@ -80,6 +85,7 @@ class App extends Model
 
     /**
      * Get the URL to edit the app in the CP
+     *
      * @return string
      */
     public function getCpEditUrl(): string
@@ -89,6 +95,7 @@ class App extends Model
 
     /**
      * Get the URL callback URL
+     *
      * @return string
      */
     public function getRedirectUrl(): string
@@ -96,7 +103,13 @@ class App extends Model
         return UrlHelper::cpUrl('oauthclient/authorize/' . $this->handle);
     }
 
-    public function getProviderInstance()
+    /**
+     * Get an instance of the Provider
+     *
+     * @return Provider|null
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function getProviderInstance(): ?Provider
     {
         if ($this->providerInstance instanceof Provider) {
             return $this->providerInstance;
@@ -110,17 +123,47 @@ class App extends Model
         return $this->providerInstance;
     }
 
-    public function getAllTokens()
+    /**
+     * Get all tokens valid tokens for the supplied user. If no user is supplied, the current user will
+     * be used.
+     *
+     * @param null|int|User $user
+     * @return Token[]
+     * @throws \Exception
+     */
+    public function getValidTokensForUser($user = null) {
+        $userId = null;
+        if ($user instanceof User) {
+            $userId = $user->id;
+        } elseif (is_int($user)) {
+            $userId = $user;
+        } elseif ($currentUser = \Craft::$app->user->getIdentity()) {
+            $userId = $currentUser->id;
+        } else {
+            // No user, but let's return an empty array so we don't break anything upstream
+            return [];
+        }
+        return Plugin::$plugin->credentials->getValidTokensForAppAndUser($this, $userId);
+    }
+
+    /**
+     * Get all token models belong to this app
+     *
+     * @return Token[]
+     */
+    public function getAllTokens(): array
     {
         return Plugin::$plugin->tokens->getAllTokensForApp($this->id);
     }
 
     /**
+     * Gets an ActiveQuery for tokens belonging to this app.
+     *
      * @return \yii\db\ActiveQuery
      */
-    public function getTokenRecordQuery()
+    public function getTokenRecordQuery(): \yii\db\ActiveQuery
     {
-        return \venveo\oauthclient\records\Token::find()->where(['appId' => $this->id]);
+        return TokenRecord::find()->where(['appId' => $this->id]);
     }
 
     /**