+++ src/providers/Google.php
@@ -10,15 +10,18 @@
 
 namespace venveo\oauthclient\providers;
 
+use GuzzleHttp\Exception\ClientException;
 use League\OAuth2\Client\Provider\Google as GoogleProvider;
 use venveo\oauthclient\base\Provider;
+use venveo\oauthclient\base\ValidatesTokens;
+use venveo\oauthclient\models\Token;
 
 /**
  * @author    Venveo
  * @package   OauthClient
  * @since     1.0.0
  */
-class Google extends Provider
+class Google extends Provider implements ValidatesTokens
 {
     /**
      * @inheritDoc
@@ -34,5 +37,21 @@ class Google extends Provider
     public static function getProviderClass(): string
     {
         return GoogleProvider::class;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public static function checkToken(Token $token): bool
+    {
+        $url = 'https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=' . $token->accessToken;
+        $client = new \GuzzleHttp\Client();
+        try {
+            $resp = $client->request('GET', $url);
+            $status = $resp->getStatusCode();
+            return ($status === 200);
+        } catch (ClientException $e) {
+            return false;
+        }
     }
 }