+++ src/services/Credentials.php
@@ -10,11 +10,17 @@
 
 namespace venveo\oauthclient\services;
 
+use Craft;
 use craft\base\Component;
 use craft\elements\User;
+use venveo\oauthclient\base\Provider;
+use venveo\oauthclient\base\ValidatesTokens;
+use venveo\oauthclient\events\TokenEvent;
 use venveo\oauthclient\models\App as AppModel;
+use venveo\oauthclient\models\Token;
 use venveo\oauthclient\models\Token as TokenModel;
 use venveo\oauthclient\Plugin;
+use yii\db\Exception;
 
 /**
  * @author    Venveo
@@ -24,29 +30,37 @@ use venveo\oauthclient\Plugin;
  */
 class Credentials extends Component
 {
+    public const EVENT_BEFORE_REFRESH_TOKEN = 'EVENT_BEFORE_REFRESH_TOKEN';
+    public const EVENT_AFTER_REFRESH_TOKEN = 'EVENT_BEFORE_REFRESH_TOKEN';
+    public const EVENT_TOKEN_REFRESH_FAILED = 'EVENT_TOKEN_REFRESH_FAILED';
+
     /**
      * Gets valid tokens given an application and optionally, a Craft user ID
      * This method will attempt to refresh expired tokens for an app
-     * @param $appHandle
-     * @param $userId
+     * @param $appHandle AppModel|string
+     * @param $user User|int
      * @return TokenModel[]
      * @throws \Exception
      */
-    public function getValidTokensForAppAndUser($appHandle, $userId = null): ?array
+    public function getValidTokensForAppAndUser($appHandle, $user = null): array
     {
         // Allow models or IDs against my better judgement
         if ($appHandle instanceof AppModel) {
             $app = $appHandle;
-            $appHandle = $app->handle;
         } else {
             $app = Plugin::$plugin->apps->getAppByHandle($appHandle);
-        }
-        if ($userId instanceof User) {
-            $userId = $userId->getId();
         }
 
         if (!$app instanceof AppModel) {
             throw new \Exception('App does not exist');
+        }
+
+        $userId = null;
+
+        if ($user instanceof User) {
+            $userId = $user->getId();
+        } elseif (is_int($user)) {
+            $userId = $user;
         }
 
         $query = $app->getTokenRecordQuery();
@@ -67,7 +81,7 @@ class Credentials extends Component
                 continue;
             }
             if ($tokenModel->hasExpired() && !$this->refreshToken($tokenModel)) {
-                \Craft::error('Unable to refresh token: '.print_r($tokenModel, true), __METHOD__);
+                Craft::error('Unable to refresh token: '.print_r($tokenModel, true), __METHOD__);
                 continue;
             }
 
@@ -81,21 +95,48 @@ class Credentials extends Component
     /**
      * Attempts to refresh a token and save it to the database
      * @param TokenModel $tokenModel
-     * @param null $app
      * @return bool
-     * @throws \League\OAuth2\Client\Provider\Exception\IdentityProviderException
      */
-    public function refreshToken(TokenModel $tokenModel, $app = null): bool
+    public function refreshToken(TokenModel $tokenModel): bool
     {
+        $event = new TokenEvent($tokenModel);
+        $this->trigger(self::EVENT_BEFORE_REFRESH_TOKEN, $event);
+
         if (!$tokenModel->refreshToken) {
             return false;
         }
-        if (!$app instanceof AppModel) {
+        try {
             $app = $tokenModel->getApp();
+            $app->getProviderInstance()->refreshToken($tokenModel);
+            $saved = Plugin::$plugin->tokens->saveToken($tokenModel);
+            if ($saved) {
+                $this->trigger(self::EVENT_AFTER_REFRESH_TOKEN, $event);
+            } else {
+                throw new Exception('Failed to save refreshed token');
+            }
+            return $saved;
+        } catch (\Exception $exception) {
+            Craft::warning($exception->getMessage(), __METHOD__);
+            $this->trigger(self::EVENT_TOKEN_REFRESH_FAILED, $event);
+            return false;
         }
+    }
 
-        $app->getProviderInstance()->refreshToken($tokenModel);
+    /**
+     * If the token's Provider implements ValidatesToken, we can ask it to verify the token with the provider
+     *
+     * @param TokenModel $token
+     * @return bool
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function checkTokenWithProvider(Token $token) {
+        $app = $token->getApp();
 
-        return Plugin::$plugin->tokens->saveToken($tokenModel);
+        /** @var Provider $provider */
+        $provider = $app->getProviderInstance();
+        if (!$provider instanceof ValidatesTokens) {
+            throw new \Exception('Provider cannot validate tokens');
+        }
+        return $provider::checkToken($token);
     }
 }